requirements.txt
Django==4.2.8
coreapi==2.3.3
python-dotenv==1.0.0
djangorestframework==3.14.0
django-cors-headers==4.3.1
psycopg2-binary==2.9.9